Only protect referenced namespaces
